+++ src/allin1/models/dinat.py
@@ -7,7 +7,7 @@ import math
 import torch
 from abc import ABC,  abstractmethod
 from typing import Optional, Tuple, Callable
-from natten.functional import natten1dav, natten1dqkrpb, natten2dav, natten2dqkrpb
+from .natten_compat import na1d_qk, na1d_av, na2d_qk, na2d_av
 from ..config import Config
 from .utils import *
 
@@ -95,8 +95,7 @@ class _NeighborhoodAttentionNd(ABC, nn.Module):
     query_layer = query_layer / math.sqrt(self.attention_head_size)
     
     # Compute NA between "query" and "key" to get the raw attention scores, and add relative positional biases.
-    # attention_scores = natten2dqkrpb(query_layer, key_layer, self.rpb, self.dilation)
-    attention_scores = self.nattendqkrpb(query_layer, key_layer, self.rpb, self.kernel_size, self.dilation)
+    attention_scores = self.nattendqkrpb(query_layer, key_layer, self.kernel_size, self.dilation, rpb=self.rpb)
     
     # Normalize the attention scores to probabilities.
     attention_probs = nn.functional.softmax(attention_scores, dim=-1)
@@ -105,7 +104,6 @@ class _NeighborhoodAttentionNd(ABC, nn.Module):
     # seem a bit unusual, but is taken from the original Transformer paper.
     attention_probs = self.dropout(attention_probs)
     
-    # context_layer = natten2dav(attention_probs, value_layer, self.dilation)
     context_layer = self.nattendav(attention_probs, value_layer, self.kernel_size, self.dilation)
     if len(context_layer.shape) > 4:  # 2D
       context_layer = context_layer.permute(0, 2, 3, 1, 4).contiguous()
@@ -141,8 +139,8 @@ class NeighborhoodAttention1d(_NeighborhoodAttentionNd):
       torch.zeros(num_heads, (2 * self.kernel_size - 1)),
       requires_grad=True,
     )
-    self.nattendqkrpb = natten1dqkrpb
-    self.nattendav = natten1dav
+    self.nattendqkrpb = na1d_qk
+    self.nattendav = na1d_av
 
 
 class NeighborhoodAttention2d(_NeighborhoodAttentionNd):
@@ -159,8 +157,8 @@ class NeighborhoodAttention2d(_NeighborhoodAttentionNd):
       torch.zeros(num_heads, (2 * self.kernel_size - 1), (2 * self.kernel_size - 1)),
       requires_grad=True,
     )
-    self.nattendqkrpb = natten2dqkrpb
-    self.nattendav = natten2dav
+    self.nattendqkrpb = na2d_qk
+    self.nattendav = na2d_av
 
 
 # Copied from transformers.models.nat.modeling_nat.NeighborhoodAttentionOutput